+++ src/components/faq_section/faq_section.styl
@@ -9,18 +9,6 @@
     top 10px
     position absolute
 
-  .anchor
-    position absolute
-    right 15%
-    bottom 40px
-    transform rotate(30deg)
-    width 70px
-    opacity 0.4
-
-    @media mediaPhone
-      bottom 5px
-      width 30px
-
   .stones
     bottom -7px
     width 102%
