+++ src/constants.styl
@@ -4,6 +4,8 @@
 
 colorPink = rgb(235, 30, 122)
 colorPinkLight = rgb(239, 83, 82)
+colorBlueDarkest = rgb(39, 58, 70)
+colorBlueDarker = rgb(19, 85, 118)
 colorBlueDark = rgb(0, 129, 189)
 colorBlue = rgb(0, 178, 226)
 colorBlueLight = rgb(44, 175, 230)
